requirements.txt
# Business Assistant GO - Requirements
# Python 3.12+

# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
psycopg2-binary==2.9.9

# HTTP Requests
requests==2.31.0

# Environment Variables
python-dotenv==1.0.0

# Twilio (для WhatsApp)
twilio==8.12.0

# Production Server
gunicorn==21.2.0

# Development & Testing
pytest==8.0.0
pytest-flask==1.3.0
